Don't auto-create an admin user
After running comprod2 create, you have to manually add (or invite) the
first user. Any (or no) user can be set to admin after the fact.

By separating admin from the command line and storing it in the database,
the game admin can be changed without having to visit the system startup
scripts.